add queries to comment subgraph

+++ src/main/scala/muse/server/graphql/resolver/GetComment.scala
@@ -8,11 +8,12 @@ import java.sql.SQLException
 import java.util.UUID
 
 object GetComment {
-  case class CommentRequest(id: Int) extends Request[SQLException, Option[Comment]]
+  case class CommentRequest(id: Long) extends Request[SQLException, Option[Comment]]
 
-  def query(id: Int): ZQuery[DatabaseService, SQLException, Option[Comment]] =
+  def query(id: Long): ZQuery[DatabaseService, SQLException, Option[Comment]] =
     ZQuery.fromRequest(CommentRequest(id))(commentDataSource)
 
+  // TODO: make this batched data source.
   val commentDataSource: DataSource[DatabaseService, CommentRequest] =
     DataSource.fromFunctionZIO("GetCommentEntities")(req =>
       DatabaseService.getComment(req.id).map { maybeCommentTuple =>
